refactor(control-ui): split index.tsx into focused modules
Move index.tsx's mixed theme/component/hook/logic pieces into dedicated
modules, one per concern:

- globalStyle.tsx: theme tokens, GlobalStyle, ThemeToggle
- hotkeyLabel.ts: hotkeyTriggers/hotkeyLayers, getHotkeyLabel
- GridPreviewBox.tsx (+ colorTypes.ts, OrientationIndicator.tsx): the
  per-tile preview box and its styled deps
- GridInput.tsx: the per-tile stream-id input
- GridControls.tsx (+ StyledButton.tsx): the per-tile control buttons,
  plus the generic button primitives shared with the header/streamdelay
  box
- GridSizeControls.tsx: the column/row presets and inputs
- Sidebar.tsx: StreamLine, StreamList, CustomStreamInput,
  CreateCustomStreamInput
- AccessPanel.tsx: CreateInviteInput, AuthTokenLine
- useTileDrag.ts / useTileResize.ts: the swap-click, drag-move, and
  tile-resize (pointer + keyboard) gesture state that was ~270 lines of
  hooks inline in ControlUI. useTileResize takes hoveringIdx from
  useTileDrag as an input; each hook now owns its own
  escape-cancels-this-gesture hotkey internally instead of ControlUI
  clearing both from one combined handler.

index.tsx shrinks from 3,092 to 1,504 lines. It re-exports only the
package's actual public API (ControlUI, GlobalStyle, CollabData,
StreamwallConnection, useStreamwallState, useYDoc — verified against
every consumer in streamwall-control-client and streamwall's renderer).

Repointed every pre-existing scaffolded test file (GridControls,
GridInput, GridPreviewBox, globalStyle, hotkeyLabel) at its new module,
preserving the upstream test additions (aria-label/tabIndex coverage,
chorded-hotkey cases) each had picked up in the meantime. Added new
coverage for modules that had none: GridSizeControls, Sidebar,
AccessPanel, useTileDrag, useTileResize (including the keyboard-resize
path) — using small harness components + a real Y.Doc for the two
hooks, since neither Preact nor this repo has a renderHook-style
utility.

Closes #44
