+++ packages/streamwall-control-ui/src/index.tsx
@@ -7,89 +7,65 @@ import '@fontsource/jetbrains-mono/500.css'
 import '@fontsource/jetbrains-mono/700.css'
 import '@fontsource/oswald/600.css'
 import '@fontsource/saira-stencil-one'
-import { orderBy, range, truncate } from 'lodash-es'
+import { orderBy, range } from 'lodash-es'
 import { DateTime } from 'luxon'
 import { type JSX } from 'preact'
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useState,
-} from 'preact/hooks'
+import { useCallback, useEffect, useMemo, useState } from 'preact/hooks'
 import { useHotkeys } from 'react-hotkeys-hook'
 import {
-  FaExchangeAlt,
-  FaExclamationTriangle,
-  FaRedoAlt,
-  FaRegLifeRing,
-  FaRegWindowMaximize,
-  FaSyncAlt,
-  FaVideoSlash,
-  FaVolumeUp,
-} from 'react-icons/fa'
-import {
-  MdOutlineStayCurrentLandscape,
-  MdOutlineStayCurrentPortrait,
-} from 'react-icons/md'
-import {
   clampGridDimension,
-  Color,
-  type ContentKind,
   type ControlCommand,
   type DataSourceHealth,
   type DisconnectReason,
-  GRID_MAX,
-  GRID_MIN,
   gridWouldDropAssignments,
   hasGridAssignments,
   idColor,
   inviteLink,
   type LayoutPreset,
   type LocalStreamData,
-  parseGridDimensionInput,
   roleCan,
   type StreamData,
   type StreamDelayStatus,
   type StreamwallRole,
   type StreamwallState,
   type StreamWindowConfig,
-  type ViewPos,
   type ViewState,
 } from 'streamwall-shared'
-import { createGlobalStyle, styled } from 'styled-components'
+import { styled } from 'styled-components'
 import { matchesState } from 'xstate'
 import * as Y from 'yjs'
+import { AuthTokenLine, CreateInviteInput } from './AccessPanel.tsx'
 import { copyTextToClipboard } from './clipboard.ts'
 import { createErrorSurfacingSend } from './commandError.ts'
 import { CommandErrorBanner } from './CommandErrorBanner.tsx'
 import { ConnectionStatusBanner } from './ConnectionStatusBanner.tsx'
 import { DataSourceHealthBanner } from './DataSourceHealthBanner.tsx'
-import {
-  computeHoveringIdx,
-  isPrimaryButton,
-  resolveMoveTarget,
-} from './gestures'
-import {
-  computeKeyboardResizeHoverIdx,
-  computeResizeAssignments,
-  computeSwap,
-  isIdxInResizeBox,
-  type ResizeHandle,
-  type SwapBox,
-} from './gridInteractions'
+import { ThemeToggle } from './globalStyle.tsx'
+import { GridControls } from './GridControls.tsx'
+import { GridInput } from './GridInput.tsx'
+import { isIdxInResizeBox } from './gridInteractions'
+import { GridPreviewBox } from './GridPreviewBox.tsx'
+import { GridSizeControls } from './GridSizeControls.tsx'
+import { hotkeyLayerBindings, hotkeyTriggers } from './hotkeyLabel.ts'
 import './index.css'
 import { LayoutPresetControls } from './LayoutPresetControls.tsx'
-import { LazyChangeInput } from './LazyChangeInput.tsx'
+import {
+  CreateCustomStreamInput,
+  CustomStreamInput,
+  StreamList,
+} from './Sidebar.tsx'
+import { StyledButton, StyledDataContainer } from './StyledButton.tsx'
+import { useTileDrag } from './useTileDrag.ts'
+import { useTileResize } from './useTileResize.ts'
 import {
   resolveEagerWriteStreamId,
   resolveTargetViewIdx,
 } from './viewPlacement.ts'
 
-// `import { Color } from 'streamwall-shared'` binds only the value; alias
-// the instance type (as returned by the Color factory) for use in
-// styled-component prop types.
-type ColorInstance = ReturnType<typeof Color>
+// Re-exported for `streamwall-control-client` and `streamwall`'s renderer,
+// which mount it alongside `<ControlUI>` — its implementation now lives in
+// `./globalStyle.tsx` alongside the theme tokens it applies.
+export { GlobalStyle } from './globalStyle.tsx'
 
 export interface ViewInfo {
   state: ViewState
@@ -104,363 +80,6 @@ interface Invite {
   tokenId: string
   name: string
   secret: string
-}
-
-const hotkeyTriggers = [
-  '1',
-  '2',
-  '3',
-  '4',
-  '5',
-  '6',
-  '7',
-  '8',
-  '9',
-  '0',
-  'q',
-  'w',
-  'e',
-  'r',
-  't',
-  'y',
-  'u',
-  'i',
-  'o',
-  'p',
-]
-
-/**
- * Audio-listen hotkeys are laid out in "layers": each layer maps the same 20
- * trigger keys to a block of grid cells, distinguished by an extra modifier.
- * Layer 0 (`alt+<key>`) covers cells 0-19; layer 1 (`alt+ctrl+<key>`) covers
- * cells 20-39. `alt+shift+<key>` is already taken by the blur toggle, so the
- * second audio layer stacks `ctrl` instead. (Caveat: on Windows international
- * layouts `Ctrl+Alt` acts as AltGr; acceptable since the control UI runs
- * primarily in the Electron control window and the handler preventDefault()s.)
- *
- * Grids can have up to GRID_MAX * GRID_MAX = 64 cells, but cells 40-63 (only
- * reachable on 7x7/8x8 grids) are intentionally left without a hotkey rather
- * than introducing a fragile triple-modifier layer that conflicts with OS
- * shortcuts.
- */
-const hotkeyLayers = [
-  { prefix: 'alt', label: 'Alt' },
-  { prefix: 'alt+ctrl', label: 'Alt+Ctrl' },
-] as const
-
-/** The `alt+<...>+<key>` combos each layer binds, e.g. `alt+1,alt+2,...`. */
-const hotkeyLayerBindings = hotkeyLayers.map((layer) =>
-  hotkeyTriggers.map((k) => `${layer.prefix}+${k}`).join(','),
-)
-
-/**
- * Label for the audio-toggle hotkey assigned to grid cell `idx` (e.g. `Alt+1`
- * or `Alt+Ctrl+1`), or `undefined` if `idx` falls outside every hotkey layer.
- */
-export function getHotkeyLabel(idx: number): string | undefined {
-  if (idx < 0) {
-    return undefined
-  }
-  const layer = hotkeyLayers[Math.floor(idx / hotkeyTriggers.length)]
-  const key = hotkeyTriggers[idx % hotkeyTriggers.length]
-  if (layer === undefined || key === undefined) {
-    return undefined
-  }
-  return `${layer.label}+${key.toUpperCase()}`
-}
-
-/**
- * Theme tokens. Light is the base; dark is applied either by the OS setting
- * (prefers-color-scheme) when no explicit choice is made, or by an explicit
- * `data-theme` attribute on <html> (set by the theme switcher).
- *
- *   <html>                       -> follow OS
- *   <html data-theme="system">   -> follow OS
- *   <html data-theme="light">    -> force light
- *   <html data-theme="dark">     -> force dark
- */
-const lightTokens = `
-  --bg:         #eef1f6;
-  --surface:    #ffffff;
-  --surface-2:  #f4f6fa;
-  --surface-3:  #e9edf3;
-  --border:     #dce1ea;
-  --border-2:   #c9d1dd;
-  --text:       #161b24;
-  --text-dim:   #5a6473;
-  --text-faint: #9aa3b1;
-  --accent:     #d6402a;
-  --accent-2:   #0a84ff;
-  --accent-soft:#fbe4df;
-  --live:       #d6402a;
-  --ok:         #17a35a;
-  --cell-bg:    #20242c;
-  --shadow:     0 6px 22px rgba(40,55,80,.12);
-`
-const darkTokens = `
-  --bg:         #0b0d11;
-  --surface:    #13161c;
-  --surface-2:  #191d25;
-  --surface-3:  #222731;
-  --border:     #2a303b;
-  --border-2:   #353c49;
-  --text:       #e8ecf2;
-  --text-dim:   #939cab;
-  --text-faint: #5b6470;
-  --accent:     #f24d2e;
-  --accent-2:   #4cc2ff;
-  --accent-soft:#2a1a16;
-  --live:       #ff445e;
-  --ok:         #3ddc84;
-  --cell-bg:    #0e1115;
-  --shadow:     0 8px 28px rgba(0,0,0,.45);
-`
-
-export const GlobalStyle = createGlobalStyle`
-  :root {
-    color-scheme: light dark;
-
-    --font-display: 'Saira Stencil One', 'Oswald', system-ui, sans-serif;
-    --font-ui: 'IBM Plex Sans', system-ui, sans-serif;
-    --font-mono: 'JetBrains Mono', ui-monospace, 'SF Mono', monospace;
-
-    --r-sm: 6px;
-    --r-md: 9px;
-    --r-lg: 12px;
-    --sp: 4px;
-
-    /* base = light */
-    ${lightTokens}
-  }
-
-  /* Follow the OS when no explicit choice was made */
-  @media (prefers-color-scheme: dark) {
-    :root:not([data-theme]),
-    :root[data-theme='system'] {
-      ${darkTokens}
-    }
-  }
-
-  /* Explicit overrides from the theme switcher */
-  :root[data-theme='light'] { ${lightTokens} }
-  :root[data-theme='dark']  { ${darkTokens} }
-
-  html {
-    height: 100%;
-  }
-
-  html, body {
-    display: flex;
-    flex: 1;
-    margin: 0;
-    /* body is a flex item of html with no explicit min-width, so its
-       automatic minimum size falls back to its content's min-content width -
-       which can exceed the viewport and inflate body past it even with the
-       margin reset above (see #225). */
-    min-width: 0;
-    background: var(--bg);
-    color: var(--text);
-    font-family: var(--font-ui);
-  }
-
-  * { box-sizing: border-box; }
-
-  /* ---- Sidebar (stream list / custom streams / access) ---- */
-  .stream-list h2 {
-    font-family: var(--font-display);
-    font-weight: normal;
-    font-size: 15px;
-    letter-spacing: 0.04em;
-    color: var(--text);
-    margin: 24px 0 12px;
-    padding-bottom: 6px;
-    border-bottom: 2px solid var(--accent);
-    display: inline-block;
-  }
-  .stream-list h3 {
-    font-size: 10.5px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.16em;
-    color: var(--text-dim);
-    margin: 18px 0 8px;
-  }
-  .stream-list h3 .ct {
-    font-family: var(--font-mono);
-    font-size: 10px;
-    color: var(--text-faint);
-    background: var(--surface-2);
-    border: 1px solid var(--border);
-    border-radius: 999px;
-    padding: 1px 7px;
-    margin-left: 6px;
-  }
-  .stream-list input,
-  .stream-list select {
-    font-family: var(--font-ui);
-    font-size: 13px;
-    color: var(--text);
-    background: var(--surface-2);
-    border: 1px solid var(--border);
-    border-radius: var(--r-sm);
-    padding: 7px 10px;
-  }
-  .stream-list input::placeholder { color: var(--text-faint); }
-  .stream-list input:focus,
-  .stream-list select:focus {
-    outline: none;
-    border-color: var(--accent);
-    box-shadow: 0 0 0 3px var(--accent-soft);
-  }
-  .stream-list .filter-input {
-    width: 100%;
-    margin-bottom: 4px;
-  }
-  .stream-list button {
-    font-family: var(--font-ui);
-    font-size: 12px;
-    font-weight: 600;
-    color: #fff;
-    background: var(--accent);
-    border: 0;
-    border-radius: var(--r-sm);
-    padding: 7px 12px;
-    cursor: pointer;
-  }
-  .stream-list button:hover { filter: brightness(1.08); }
-  .stream-list a { color: var(--accent-2); text-decoration: none; }
-  .stream-list a:hover { text-decoration: underline; }
-`
-
-type ThemeChoice = 'system' | 'light' | 'dark'
-const THEME_KEY = 'streamwall:theme'
-
-const StyledThemeToggle = styled.div`
-  display: inline-flex;
-  gap: 2px;
-  padding: 3px;
-  border-radius: var(--r-sm);
-  background: var(--surface-2);
-  border: 1px solid var(--border);
-
-  button {
-    display: grid;
-    place-items: center;
-    width: 27px;
-    height: 24px;
-    border: 0;
-    border-radius: 4px;
-    background: transparent;
-    color: var(--text-faint);
-    cursor: pointer;
-    padding: 0;
-  }
-  button:hover {
-    color: var(--text-dim);
-  }
-  button.active {
-    background: var(--accent);
-    color: #fff;
-  }
-  svg {
-    width: 15px;
-    height: 15px;
-  }
-`
-
-/**
- * Theme switcher. Writes the choice to <html data-theme> (read by GlobalStyle)
- * and persists it in localStorage. 'system' clears the override so the OS
- * setting (prefers-color-scheme) takes over.
- */
-export function ThemeToggle() {
-  const [theme, setTheme] = useState<ThemeChoice>(() => {
-    try {
-      return (localStorage.getItem(THEME_KEY) as ThemeChoice) ?? 'system'
-    } catch {
-      return 'system'
-    }
-  })
-
-  useEffect(() => {
-    document.documentElement.dataset.theme = theme
-    try {
-      localStorage.setItem(THEME_KEY, theme)
-    } catch {
-      // ignore (e.g. storage disabled)
-    }
-  }, [theme])
-
-  const opts = [
-    {
-      key: 'system' as const,
-      label: 'System',
-      icon: (
-        <svg
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          stroke-width="2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <rect x="2" y="3" width="20" height="14" rx="2" />
-          <path d="M8 21h8M12 17v4" />
-        </svg>
-      ),
-    },
-    {
-      key: 'light' as const,
-      label: 'Hell',
-      icon: (
-        <svg
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          stroke-width="2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <circle cx="12" cy="12" r="4" />
-          <path d="M12 2v2M12 20v2M4.9 4.9l1.4 1.4M17.7 17.7l1.4 1.4M2 12h2M20 12h2M4.9 19.1l1.4-1.4M17.7 6.3l1.4-1.4" />
-        </svg>
-      ),
-    },
-    {
-      key: 'dark' as const,
-      label: 'Dunkel',
-      icon: (
-        <svg
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          stroke-width="2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        >
-          <path d="M21 12.8A9 9 0 1 1 11.2 3a7 7 0 0 0 9.8 9.8z" />
-        </svg>
-      ),
-    },
-  ]
-
-  return (
-    <StyledThemeToggle role="group" aria-label="Color scheme">
-      {opts.map(({ key, label, icon }) => (
-        <button
-          key={key}
-          type="button"
-          class={theme === key ? 'active' : undefined}
-          title={label}
-          aria-label={label}
-          aria-pressed={theme === key}
-          onClick={() => setTheme(key)}
-        >
-          {icon}
-        </button>
-      ))}
-    </StyledThemeToggle>
-  )
 }
 
 const normalStreamKinds = new Set(['video', 'audio', 'web'])
@@ -651,279 +270,23 @@ export function ControlUI({
     setShowDebug(ev.currentTarget.checked)
   }, [])
 
-  const [swapStartIdx, setSwapStartIdx] = useState<number | undefined>()
-  const handleSwapView = useCallback(
-    (idx: number) => {
-      if (!stateIdxMap.has(idx)) {
-        return
-      }
-      // Deselect the input so the contents aren't persisted by GridInput's `editingValue`
-      const { activeElement } = document
-      if (activeElement && activeElement instanceof HTMLElement) {
-        activeElement.blur()
-      }
-      setSwapStartIdx(idx)
-    },
-    [stateIdxMap],
-  )
-  // Used by both the swap gesture (`handleSwap`) and the drag-move gesture
-  // (`endMove` below) to commit a two-box exchange.
-  //
-  // Known limitation: this reads each box's current streamId and writes both
-  // boxes' new streamIds as independent keys in the shared Yjs map. Yjs
-  // resolves concurrent edits per key (last-writer-wins), not across the pair
-  // atomically. If two operators swap/move overlapping boxes at nearly the
-  // same time, each computes its reassignment from a state the other is also
-  // about to overwrite — the merged result can duplicate one stream into both
-  // boxes while dropping the other. This is inherent to modeling the grid as
-  // independent per-cell keys rather than a single atomically-swapped value;
-  // a real fix would need a server-side or CRDT-level guard (e.g. rejecting a
-  // swap whose read boxes changed since it started). Low impact in practice
-  // since concurrent operators rarely target the same cells simultaneously.
-  const swapBoxes = useCallback(
-    (fromIdx: number, toIdx: number) => {
-      if (cols == null || rows == null) {
-        return
-      }
-      stateDoc.transact(() => {
-        const viewsMap = stateDoc.getMap<Y.Map<string | undefined>>('views')
-        const boxes = new Map<number, SwapBox>()
-        for (const idx of [fromIdx, toIdx]) {
-          const viewInfo = stateIdxMap.get(idx)
-          if (viewInfo === undefined) {
-            // No box occupies this index — leave it out of `boxes` so
-            // `computeSwap` treats it as a genuinely empty target rather
-            // than a single-space box, letting it translate a merged
-            // source box's whole footprint there instead of collapsing it.
-            continue
-          }
-          boxes.set(idx, {
-            spaces: viewInfo.spaces,
-            streamId: viewsMap.get(String(idx))?.get('streamId'),
-          })
-        }
-        const assignments = computeSwap(boxes, fromIdx, toIdx, cols, rows)
-        for (const [idx, streamId] of assignments) {
-          viewsMap.get(String(idx))?.set('streamId', streamId)
-        }
-      })
-    },
-    [stateDoc, stateIdxMap, cols, rows],
-  )
-
-  const handleSwap = useCallback(
-    (toIdx: number) => {
-      if (swapStartIdx === undefined) {
-        return
-      }
-      swapBoxes(swapStartIdx, toIdx)
-      setSwapStartIdx(undefined)
-    },
-    [swapBoxes, swapStartIdx],
-  )
-
-  const [hoveringIdx, setHoveringIdx] = useState<number>()
-  const updateHoveringIdx = useCallback(
-    (ev: PointerEvent) => {
-      if (
-        cols == null ||
-        rows == null ||
-        !(ev.currentTarget instanceof HTMLElement)
-      ) {
-        return
-      }
-      const { width, height, left, top } =
-        ev.currentTarget.getBoundingClientRect()
-      setHoveringIdx(
-        computeHoveringIdx(
-          cols,
-          rows,
-          width,
-          height,
-          ev.clientX - left,
-          ev.clientY - top,
-        ),
-      )
-    },
-    [setHoveringIdx, cols, rows],
-  )
-  // Clear the hovered cell when the pointer leaves the grid so a gesture
-  // released off-grid can't commit against a stale cell. This only fires for
-  // mouse pointers — a touch pointer is implicitly captured to its
-  // `pointerdown` target and never dispatches boundary events while
-  // dragging, so `updateHoveringIdx`'s own bounds check covers that case.
-  const clearHoveringIdx = useCallback(() => setHoveringIdx(undefined), [])
-  const [moveStart, setMoveStart] = useState<
-    { idx: number; x: number; y: number } | undefined
-  >()
-  const [moveTargetIdx, setMoveTargetIdx] = useState<number | undefined>()
-
-  const handleGridPointerDown = useCallback(
-    (ev: PointerEvent) => {
-      if (!isPrimaryButton(ev.button) || hoveringIdx == null) {
-        return
-      }
-      if (swapStartIdx !== undefined) {
-        handleSwap(hoveringIdx)
-        return
-      }
-      setMoveStart({ idx: hoveringIdx, x: ev.clientX, y: ev.clientY })
-    },
-    [hoveringIdx, swapStartIdx, handleSwap],
-  )
-
-  useLayoutEffect(() => {
-    if (moveStart == null) {
-      setMoveTargetIdx(undefined)
-      return
-    }
-    setMoveTargetIdx(hoveringIdx)
-  }, [moveStart, hoveringIdx])
-
-  useLayoutEffect(() => {
-    function endMove(ev: PointerEvent) {
-      if (moveStart == null) {
-        return
-      }
-      // pointercancel means the gesture was interrupted (e.g. a system
-      // gesture taking over on touch) — abort without committing.
-      if (ev.type !== 'pointercancel') {
-        const targetIdx = resolveMoveTarget(
-          moveStart,
-          hoveringIdx,
-          ev.clientX,
-          ev.clientY,
-        )
-        if (targetIdx != null) {
-          swapBoxes(moveStart.idx, targetIdx)
-        }
-      }
-      setMoveStart(undefined)
-    }
-    window.addEventListener('pointerup', endMove)
-    window.addEventListener('pointercancel', endMove)
-    return () => {
-      window.removeEventListener('pointerup', endMove)
-      window.removeEventListener('pointercancel', endMove)
-    }
-  }, [moveStart, hoveringIdx, swapBoxes])
-
-  const [resize, setResize] = useState<
-    | {
-        anchorIdx: number
-        streamId: string
-        handle: ResizeHandle
-        originalSpaces: number[]
-      }
-    | undefined
-  >()
-
-  const handleResizeStart = useCallback(
-    (
-      anchorIdx: number,
-      handle: ResizeHandle,
-      originalSpaces: number[],
-      ev: PointerEvent,
-    ) => {
-      if (!isPrimaryButton(ev.button)) {
-        return
-      }
-      ev.preventDefault()
-      ev.stopPropagation()
-      const streamId = sharedState?.views?.[anchorIdx]?.streamId ?? undefined
-      if (streamId == null || streamId === '') {
-        return
-      }
-      setResize({ anchorIdx, streamId, handle, originalSpaces })
-    },
-    [sharedState],
-  )
-
-  // Keyboard equivalent of the pointer-drag resize above: each arrow-key
-  // press commits a one-cell step immediately (there's no keyboard hover
-  // state to preview), rather than opening an in-progress `resize` gesture.
-  const handleResizeKeyDown = useCallback(
-    (
-      anchorIdx: number,
-      handle: ResizeHandle,
-      originalSpaces: number[],
-      ev: KeyboardEvent,
-    ) => {
-      if (cols == null || rows == null) {
-        return
-      }
-      const hoverIdx = computeKeyboardResizeHoverIdx(
-        cols,
-        rows,
-        anchorIdx,
-        handle,
-        originalSpaces,
-        ev.key,
-      )
-      if (hoverIdx == null) {
-        return
-      }
-      ev.preventDefault()
-      const streamId = sharedState?.views?.[anchorIdx]?.streamId ?? undefined
-      if (streamId == null || streamId === '') {
-        return
-      }
-      stateDoc.transact(() => {
-        const viewsMap = stateDoc.getMap<Y.Map<string | undefined>>('views')
-        const assignments = computeResizeAssignments(
-          cols,
-          anchorIdx,
-          hoverIdx,
-          streamId,
-          handle,
-          originalSpaces,
-        )
-        for (const [idx, assignedStreamId] of assignments) {
-          viewsMap.get(String(idx))?.set('streamId', assignedStreamId)
-        }
-      })
-    },
-    [cols, rows, sharedState, stateDoc],
-  )
-
-  useLayoutEffect(() => {
-    function endResize(ev: PointerEvent) {
-      // A resize only commits while the pointer is over the grid; released
-      // off-grid `hoveringIdx` is cleared, so this aborts instead of
-      // snapping to a stale cell. A pointercancel likewise aborts.
-      if (
-        ev.type === 'pointercancel' ||
-        resize == null ||
-        cols == null ||
-        rows == null ||
-        hoveringIdx == null
-      ) {
-        setResize(undefined)
-        return
-      }
-      stateDoc.transact(() => {
-        const viewsMap = stateDoc.getMap<Y.Map<string | undefined>>('views')
-        const assignments = computeResizeAssignments(
-          cols,
-          resize.anchorIdx,
-          hoveringIdx,
-          resize.streamId,
-          resize.handle,
-          resize.originalSpaces,
-        )
-        for (const [idx, streamId] of assignments) {
-          viewsMap.get(String(idx))?.set('streamId', streamId)
-        }
-      })
-      setResize(undefined)
-    }
-    window.addEventListener('pointerup', endResize)
-    window.addEventListener('pointercancel', endResize)
-    return () => {
-      window.removeEventListener('pointerup', endResize)
-      window.removeEventListener('pointercancel', endResize)
-    }
-  }, [resize, cols, rows, hoveringIdx, stateDoc])
+  const {
+    hoveringIdx,
+    swapStartIdx,
+    moveStart,
+    moveTargetIdx,
+    updateHoveringIdx,
+    clearHoveringIdx,
+    handleSwapView,
+    handleGridPointerDown,
+  } = useTileDrag({ cols, rows, stateDoc, stateIdxMap })
+  const { resize, handleResizeStart, handleResizeKeyDown } = useTileResize({
+    cols,
+    rows,
+    hoveringIdx,
+    stateDoc,
+    sharedState,
+  })
 
   const [focusedInputIdx, setFocusedInputIdx] = useState<number | undefined>()
   const handleBlurInput = useCallback(() => setFocusedInputIdx(undefined), [])
@@ -1269,18 +632,6 @@ export function ControlUI({
       }
     },
     [handleSwapView, focusedInputIdx],
-  )
-  // Escape cancels an in-progress drag-move or resize without committing. The
-  // window pointerup/pointercancel listeners are no-ops once these are cleared.
-  useHotkeys(
-    `escape`,
-    () => {
-      setMoveStart(undefined)
-      setResize(undefined)
-    },
-    // Also fire while a grid input is focused during a gesture.
-    { enableOnFormTags: true },
-    [setMoveStart, setResize],
   )
   // Undo/redo edits to the shared views doc (drag-move, swap, and destructive
   // grid-shrink remaps alike - see `useYDoc`'s `remoteOrigin` wiring).
@@ -1877,163 +1228,6 @@ function StreamDelayBox({
   )
 }
 
-function OrientationIndicator({
-  orientation,
-  className,
-}: {
-  orientation: 'V' | 'H' | null | undefined
-  className?: string
-}) {
-  if (orientation === 'V') {
-    return (
-      <span className={className}>
-        <MdOutlineStayCurrentPortrait />
-      </span>
-    )
-  } else if (orientation === 'H') {
-    return (
-      <span className={className}>
-        <MdOutlineStayCurrentLandscape />
-      </span>
-    )
-  } else {
-    return null
-  }
-}
-
-// Extracted from the ControlUI grid preview loop so the error badge markup
-// can be rendered and tested in isolation.
-export function GridPreviewBox({
-  streamId,
-  color,
-  pos,
-  windowWidth,
-  windowHeight,
-  isListening,
-  isSmall,
-  isError,
-  errorReason,
-  orientation,
-  source,
-  city,
-  state,
-  style,
-}: {
-  streamId: string
-  color: ColorInstance
-  pos: ViewPos
-  windowWidth: number
-  windowHeight: number
-  isListening: boolean
-  isSmall: boolean
-  isError: boolean
-  errorReason: string | null | undefined
-  orientation: 'V' | 'H' | null | undefined
-  source: string | undefined
-  city: string | undefined
-  state: string | undefined
-  style?: JSX.HTMLAttributes['style']
-}) {
-  const hotkeyLabel = getHotkeyLabel(pos.spaces[0])
-  return (
-    <StyledGridPreviewBox
-      $color={color}
-      style={style}
-      $pos={pos}
-      $windowWidth={windowWidth}
-      $windowHeight={windowHeight}
-      $isListening={isListening}
-      $isError={isError}
-      title={hotkeyLabel ? `${hotkeyLabel} toggles audio` : undefined}
-    >
-      <StyledGridInfo className={isSmall ? 'small' : undefined}>
-        <StyledGridLabel>
-          {streamId}
-          <OrientationIndicator
-            orientation={orientation}
-            className={`orientation-${(orientation ?? 'unknown').toLowerCase()}`}
-          />
-        </StyledGridLabel>
-        {!isSmall && <div>{source}</div>}
-        {city && (
-          <StyledGridLocation>
-            {city} {state}
-          </StyledGridLocation>
-        )}
-        {isError && (
-          <StyledGridError title={errorReason ?? undefined}>
-            <FaExclamationTriangle />
-            <span>{errorReason ?? 'Stream error'}</span>
-          </StyledGridError>
-        )}
-      </StyledGridInfo>
-    </StyledGridPreviewBox>
-  )
-}
-
-export function StreamList({
-  rows,
-  disabled,
-  onClickId,
-}: {
-  rows: StreamData[]
-  disabled: boolean
-  onClickId: (id: string) => void
-}) {
-  return rows.map((row) => (
-    <StreamLine
-      key={row._id}
-      id={row._id}
-      row={row}
-      disabled={disabled}
-      onClickId={onClickId}
-    />
-  ))
-}
-
-function StreamLine({
-  id,
-  row: { label, source, link, notes, city, state, orientation },
-  disabled,
-  onClickId,
-}: {
-  id: string
-  row: StreamData
-  disabled: boolean
-  onClickId: (id: string) => void
-}) {
-  // Use mousedown instead of click event so a potential destination grid input stays focused.
-  const handleMouseDownId = useCallback(() => {
-    onClickId(id)
-  }, [onClickId, id])
-  return (
-    <StyledStreamLine>
-      <StyledId
-        $disabled={disabled}
-        onMouseDown={disabled ? undefined : handleMouseDownId}
-        $color={idColor(id)}
-      >
-        {id}
-      </StyledId>
-      <div>
-        {label ? (
-          label
-        ) : (
-          <>
-            <strong>{source}</strong>{' '}
-            <OrientationIndicator orientation={orientation} />{' '}
-            {city ? `(${city} ${state}) ` : ''}
-            <a href={link} target="_blank">
-              {truncate(link, { length: 55 })}
-            </a>{' '}
-            {notes}
-          </>
-        )}
-      </div>
-    </StyledStreamLine>
-  )
-}
-
 const StyledResizeHandles = styled.div`
   position: absolute;
   inset: 0;
@@ -2093,466 +1287,6 @@ const StyledResizeHandles = styled.div`
     border-radius: 2px;
   }
 `
-
-const StyledGridSizeControls = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  flex-wrap: wrap;
-
-  button.preset {
-    padding: 2px 8px;
-    border: 1px solid var(--border, #444);
-    background: transparent;
-    color: inherit;
-    border-radius: 4px;
-    cursor: pointer;
-    font: inherit;
-  }
-  button.preset.active {
-    border-color: var(--accent, #e23);
-    color: var(--accent, #e23);
-  }
-  label {
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    font-size: 12px;
-  }
-  input {
-    width: 44px;
-    background: transparent;
-    color: inherit;
-    border: 1px solid var(--border, #444);
-    border-radius: 4px;
-    padding: 2px 4px;
-  }
-`
-
-const GRID_PRESETS: Array<[number, number]> = [
-  [2, 2],
-  [3, 3],
-  [4, 4],
-  [2, 3],
-  [3, 2],
-  [4, 3],
-]
-
-function GridSizeControls({
-  cols,
-  rows,
-  role,
-  onSetGridSize,
-}: {
-  cols: number
-  rows: number
-  role: StreamwallRole | null
-  onSetGridSize: (cols: number, rows: number) => void
-}) {
-  const disabled = !roleCan(role, 'mutate-state-doc')
-
-  // The inputs hold a local draft while the user types so that transient
-  // states (an empty or out-of-range field) never reach the grid. The change
-  // is committed only on blur/Enter, and only when it parses to a valid,
-  // in-range dimension — otherwise the field reverts to the authoritative
-  // value. This prevents a cleared field (NaN) from collapsing the wall.
-  const [colsDraft, setColsDraft] = useState(String(cols))
-  const [rowsDraft, setRowsDraft] = useState(String(rows))
-  useEffect(() => setColsDraft(String(cols)), [cols])
-  useEffect(() => setRowsDraft(String(rows)), [rows])
-
-  const commitCols = useCallback(
-    (raw: string) => {
-      const parsed = parseGridDimensionInput(raw)
-      if (parsed !== null && parsed !== cols) {
-        onSetGridSize(parsed, rows)
-      }
-      setColsDraft(String(cols))
-    },
-    [cols, rows, onSetGridSize],
-  )
-  const commitRows = useCallback(
-    (raw: string) => {
-      const parsed = parseGridDimensionInput(raw)
-      if (parsed !== null && parsed !== rows) {
-        onSetGridSize(cols, parsed)
-      }
-      setRowsDraft(String(rows))
-    },
-    [cols, rows, onSetGridSize],
-  )
-
-  const commitOnEnter: JSX.KeyboardEventHandler<HTMLInputElement> = (ev) => {
-    if (ev.key === 'Enter') {
-      ev.currentTarget.blur()
-    }
-  }
-
-  return (
-    <StyledGridSizeControls>
-      {GRID_PRESETS.map(([c, r]) => (
-        <button
-          key={`${c}x${r}`}
-          type="button"
-          className={c === cols && r === rows ? 'preset active' : 'preset'}
-          disabled={disabled}
-          onClick={() => onSetGridSize(c, r)}
-        >
-          {c}×{r}
-        </button>
-      ))}
-      <label>
-        Columns
-        <input
-          type="number"
-          min={GRID_MIN}
-          max={GRID_MAX}
-          value={colsDraft}
-          disabled={disabled}
-          onInput={(ev) => setColsDraft(ev.currentTarget.value)}
-          onBlur={(ev) => commitCols(ev.currentTarget.value)}
-          onKeyDown={commitOnEnter}
-        />
-      </label>
-      <label>
-        Rows
-        <input
-          type="number"
-          min={GRID_MIN}
-          max={GRID_MAX}
-          value={rowsDraft}
-          disabled={disabled}
-          onInput={(ev) => setRowsDraft(ev.currentTarget.value)}
-          onBlur={(ev) => commitRows(ev.currentTarget.value)}
-          onKeyDown={commitOnEnter}
-        />
-      </label>
-    </StyledGridSizeControls>
-  )
-}
-
-export function GridInput({
-  style,
-  idx,
-  onChangeSpace,
-  spaceValue,
-  isHighlighted,
-  role,
-  onPointerDown,
-  onFocus,
-  onBlur,
-}: {
-  style: JSX.HTMLAttributes['style']
-  onPointerDown: JSX.PointerEventHandler<HTMLInputElement>
-  idx: number
-  onChangeSpace: (idx: number, value: string) => void
-  spaceValue: string
-  isHighlighted: boolean
-  role: StreamwallRole | null
-  onFocus: (idx: number) => void
-  onBlur: (idx: number) => void
-}) {
-  const handleFocus = useCallback(() => {
-    onFocus(idx)
-  }, [onFocus, idx])
-  const handleBlur = useCallback(() => {
-    onBlur(idx)
-  }, [onBlur, idx])
-  const handleChange = useCallback(
-    (value: string) => {
-      onChangeSpace(idx, value)
-    },
-    [idx, onChangeSpace],
-  )
-  return (
-    <StyledGridInputContainer style={style}>
-      <StyledGridInput
-        value={spaceValue}
-        $color={idColor(spaceValue)}
-        $isHighlighted={isHighlighted}
-        disabled={!roleCan(role, 'mutate-state-doc')}
-        onFocus={handleFocus}
-        onBlur={handleBlur}
-        onPointerDown={onPointerDown}
-        onChange={handleChange}
-        isEager
-      />
-    </StyledGridInputContainer>
-  )
-}
-
-export function GridControls({
-  idx,
-  streamId,
-  style,
-  isDisplaying,
-  isListening,
-  isBackgroundListening,
-  isBlurred,
-  volume,
-  isSwapping,
-  showDebug,
-  role,
-  onSetListening,
-  onSetBackgroundListening,
-  onSetBlurred,
-  onSetVolume,
-  onReloadView,
-  onSwapView,
-  onRotateView,
-  onBrowse,
-  onDevTools,
-  onPointerDown,
-}: {
-  idx: number
-  streamId: string
-  style: JSX.HTMLAttributes['style']
-  isDisplaying: boolean
-  isListening: boolean
-  isBackgroundListening: boolean
-  isBlurred: boolean
-  volume: number
-  isSwapping: boolean
-  showDebug: boolean
-  role: StreamwallRole | null
-  onSetListening: (idx: number, isListening: boolean) => void
-  onSetBackgroundListening: (
-    idx: number,
-    isBackgroundListening: boolean,
-  ) => void
-  onSetBlurred: (idx: number, isBlurred: boolean) => void
-  onSetVolume: (idx: number, volume: number) => void
-  onReloadView: (idx: number) => void
-  onSwapView: (idx: number) => void
-  onRotateView: (streamId: string) => void
-  onBrowse: (streamId: string) => void
-  onDevTools: (idx: number) => void
-  onPointerDown: JSX.PointerEventHandler<HTMLDivElement>
-}) {
-  // TODO: Refactor callbacks to use streamID instead of idx.
-  // We should probably also switch the view-state-changing RPCs to use a view id instead of idx like they do currently.
-  const handleListeningClick = useCallback<
-    JSX.MouseEventHandler<HTMLButtonElement>
-  >(
-    (ev) =>
-      ev.shiftKey || isBackgroundListening
-        ? onSetBackgroundListening(idx, !isBackgroundListening)
-        : onSetListening(idx, !isListening),
-    [
-      idx,
-      onSetListening,
-      onSetBackgroundListening,
-      isListening,
-      isBackgroundListening,
-    ],
-  )
-  const handleBlurClick = useCallback(
-    () => onSetBlurred(idx, !isBlurred),
-    [idx, onSetBlurred, isBlurred],
-  )
-  const handleVolumeInput = useCallback<
-    JSX.InputEventHandler<HTMLInputElement>
-  >(
-    (ev) => onSetVolume(idx, Number(ev.currentTarget.value)),
-    [idx, onSetVolume],
-  )
-  const handleReloadClick = useCallback(
-    () => onReloadView(idx),
-    [idx, onReloadView],
-  )
-  const handleSwapClick = useCallback(() => onSwapView(idx), [idx, onSwapView])
-  const handleRotateClick = useCallback(
-    () => onRotateView(streamId),
-    [streamId, onRotateView],
-  )
-  const handleBrowseClick = useCallback(
-    () => onBrowse(streamId),
-    [streamId, onBrowse],
-  )
-  const handleDevToolsClick = useCallback(
-    () => onDevTools(idx),
-    [idx, onDevTools],
-  )
-  return (
-    <StyledGridControlsContainer style={style} onPointerDown={onPointerDown}>
-      {isDisplaying && (
-        <StyledGridButtons $side="left">
-          {showDebug ? (
-            <>
-              {roleCan(role, 'reload-view') && (
-                <StyledSmallButton
-                  onClick={handleReloadClick}
-                  aria-label="Reload stream"
-                >
-                  <FaSyncAlt />
-                </StyledSmallButton>
-              )}
-              {roleCan(role, 'browse') && (
-                <StyledSmallButton
-                  onClick={handleBrowseClick}
-                  aria-label="Open stream in browser"
-                >
-                  <FaRegWindowMaximize />
-                </StyledSmallButton>
-              )}
-              {roleCan(role, 'dev-tools') && (
-                <StyledSmallButton
-                  onClick={handleDevToolsClick}
-                  aria-label="Open developer tools"
-                >
-                  <FaRegLifeRing />
-                </StyledSmallButton>
-              )}
-            </>
-          ) : (
-            <>
-              {roleCan(role, 'reload-view') && (
-                <StyledSmallButton
-                  onClick={handleReloadClick}
-                  aria-label="Reload stream"
-                >
-                  <FaSyncAlt />
-                </StyledSmallButton>
-              )}
-              {roleCan(role, 'mutate-state-doc') && (
-                <StyledSmallButton
-                  $isActive={isSwapping}
-                  onClick={handleSwapClick}
-                  aria-label={isSwapping ? 'Cancel swap' : 'Swap stream'}
-                >
-                  <FaExchangeAlt />
-                </StyledSmallButton>
-              )}
-              {roleCan(role, 'rotate-stream') && (
-                <StyledSmallButton
-                  onClick={handleRotateClick}
-                  aria-label="Rotate stream"
-                >
-                  <FaRedoAlt />
-                </StyledSmallButton>
-              )}
-            </>
-          )}
-        </StyledGridButtons>
-      )}
-      <StyledGridButtons $side="right">
-        {roleCan(role, 'set-view-blurred') && (
-          <StyledButton
-            $isActive={isBlurred}
-            onClick={handleBlurClick}
-            aria-label={isBlurred ? 'Unblur video' : 'Blur video'}
-          >
-            <FaVideoSlash />
-          </StyledButton>
-        )}
-        {roleCan(role, 'set-view-volume') && (
-          <StyledVolumeSlider
-            type="range"
-            min={0}
-            max={1}
-            step={0.01}
-            value={volume}
-            onInput={handleVolumeInput}
-            aria-label="Volume"
-          />
-        )}
-        {roleCan(role, 'set-listening-view') && (
-          <StyledButton
-            $isActive={isListening || isBackgroundListening}
-            $activeColor={
-              isListening ? 'red' : Color('red').desaturate(0.5).hsl().string()
-            }
-            onClick={handleListeningClick}
-            aria-label={
-              isListening || isBackgroundListening
-                ? 'Mute audio'
-                : 'Listen to audio'
-            }
-          >
-            <FaVolumeUp />
-          </StyledButton>
-        )}
-      </StyledGridButtons>
-    </StyledGridControlsContainer>
-  )
-}
-
-function CustomStreamInput({
-  onChange,
-  onDelete,
-  ...props
-}: {
-  onChange: (link: string, data: LocalStreamData) => void
-  onDelete: (link: string) => void
-} & LocalStreamData) {
-  const handleChangeLabel = useCallback(
-    (value: string) => {
-      onChange(props.link, { ...props, label: value })
-    },
-    [onChange, props],
-  )
-
-  const handleDeleteClick = useCallback(() => {
-    onDelete(props.link)
-  }, [onDelete, props.link])
-
-  return (
-    <div>
-      <LazyChangeInput
-        value={props.label ?? ''}
-        onChange={handleChangeLabel}
-        placeholder="Label (optional)"
-      />{' '}
-      <a href={props.link}>{props.link}</a> <span>({props.kind})</span>{' '}
-      <button onClick={handleDeleteClick}>x</button>
-    </div>
-  )
-}
-
-function CreateCustomStreamInput({
-  onCreate,
-}: {
-  onCreate: (link: string, data: LocalStreamData) => void
-}) {
-  const [link, setLink] = useState('')
-  const [kind, setKind] = useState<ContentKind>('video')
-  const [label, setLabel] = useState('')
-  const handleSubmit = useCallback<JSX.SubmitEventHandler<HTMLFormElement>>(
-    (ev) => {
-      ev.preventDefault()
-      onCreate(link, { link, kind, label })
-      setLink('')
-      setKind('video')
-      setLabel('')
-    },
-    [onCreate, link, kind, label],
-  )
-  return (
-    <form onSubmit={handleSubmit}>
-      <input
-        value={link}
-        onChange={(ev) => setLink(ev.currentTarget.value)}
-        placeholder="https://..."
-      />
-      <select
-        onChange={(ev) => setKind(ev.currentTarget.value as ContentKind)}
-        value={kind}
-      >
-        <option value="video">video</option>
-        <option value="audio">audio</option>
-        <option value="web">web</option>
-        <option value="overlay">overlay</option>
-        <option value="background">background</option>
-      </select>
-      <input
-        value={label}
-        onChange={(ev) => setLabel(ev.currentTarget.value)}
-        placeholder="Label (optional)"
-      />
-      <button type="submit">add stream</button>
-    </form>
-  )
-}
 
 const StyledHeader = styled.header`
   display: flex;
@@ -2659,169 +1393,12 @@ const StyledStreamDelayBox = styled.div`
   color: var(--text);
 `
 
-const StyledDataContainer = styled.div<{ $isConnected?: boolean }>`
-  opacity: ${({ $isConnected }) => ($isConnected ? 1 : 0.5)};
-`
-
-const StyledButton = styled.button<{
-  $isActive?: boolean
-  $activeColor?: string
-}>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid var(--border-2);
-  background: var(--surface-3);
-  color: var(--text-dim);
-  border-radius: var(--r-sm);
-  padding: 4px;
-  cursor: pointer;
-
-  &:hover {
-    color: var(--text);
-    border-color: var(--text-faint);
-  }
-
-  ${({ $isActive, $activeColor = 'red' }) =>
-    $isActive &&
-    `
-      border-color: ${Color($activeColor).hsl().string()};
-      background: ${Color($activeColor).hsl().string()};
-      color: #fff;
-    `};
-
-  &:focus {
-    outline: none;
-    box-shadow: 0 0 0 2px var(--accent-soft);
-  }
-
-  svg {
-    width: 20px;
-    height: 20px;
-  }
-`
-
-const StyledSmallButton = styled(StyledButton)`
-  svg {
-    width: 14px;
-    height: 14px;
-  }
-`
-
-const StyledVolumeSlider = styled.input`
-  width: 54px;
-  height: 24px;
-  margin: 5px;
-  accent-color: var(--accent-2);
-  cursor: pointer;
-`
-
 const StyledGridPreview = styled.div`
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
-`
-
-const StyledGridInfo = styled.div`
-  text-align: center;
-  padding: 8px;
-  border-radius: 16px;
-  pointer-events: none;
-  z-index: 1000; /* Keep above grid inputs */
-`
-
-const StyledGridPreviewBox = styled.div.attrs<{
-  $color: ColorInstance
-  $isError: boolean
-  $pos: ViewPos
-  $windowWidth: number
-  $windowHeight: number
-  $isListening: boolean
-  $borderWidth?: number
-}>(() => ({
-  $borderWidth: 2,
-}))`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  background: ${({ $color }) =>
-    Color($color).lightness(50).hsl().string() || '#333'};
-  border: 0 solid
-    ${({ $isError }) =>
-      $isError ? Color('red').hsl().string() : Color('black').hsl().string()};
-  border-left-width: ${({ $pos, $borderWidth }) =>
-    $pos.x === 0 ? 0 : $borderWidth}px;
-  border-right-width: ${({ $pos, $borderWidth, $windowWidth }) =>
-    $pos.x + $pos.width === $windowWidth ? 0 : $borderWidth}px;
-  border-top-width: ${({ $pos, $borderWidth }) =>
-    $pos.y === 0 ? 0 : $borderWidth}px;
-  border-bottom-width: ${({ $pos, $borderWidth, $windowHeight }) =>
-    $pos.y + $pos.height === $windowHeight ? 0 : $borderWidth}px;
-  box-shadow: ${({ $isListening }) =>
-    $isListening ? `0 0 0 4px red inset` : 'none'};
-  box-sizing: border-box;
-  overflow: hidden;
-  user-select: none;
-
-  ${StyledGridInfo} {
-    background: ${({ $color }) =>
-      Color($color).lightness(50).hsl().string() || '#333'};
-  }
-`
-
-const StyledGridLabel = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  font-size: 30px;
-
-  .orientation-v {
-    margin-left: -4px;
-  }
-
-  ${StyledGridInfo}.small & {
-    font-size: 20px;
-  }
-`
-
-const StyledGridLocation = styled.div`
-  font-size: 13px;
-  opacity: 0.75;
-`
-
-const StyledGridError = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-  margin-top: 6px;
-  padding: 3px 8px;
-  border-radius: 8px;
-  max-width: 100%;
-  font-size: 12px;
-  font-weight: 600;
-  color: white;
-  background: ${Color('red').alpha(0.7).string()};
-
-  svg {
-    flex-shrink: 0;
-  }
-
-  span {
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    overflow: hidden;
-  }
-
-  ${StyledGridInfo}.small & {
-    span {
-      display: none;
-    }
-  }
 `
 
 const StyledGridInputs = styled.div`
@@ -2834,56 +1411,6 @@ const StyledGridInputs = styled.div`
   transition: opacity 100ms ease-out;
   overflow: hidden;
   z-index: 100;
-`
-
-const StyledGridInputContainer = styled.div`
-  position: absolute;
-  touch-action: none;
-`
-
-const StyledGridButtons = styled.div<{ $side?: 'left' | 'right' }>`
-  display: flex;
-  position: absolute;
-  ${({ $side }) =>
-    $side === 'left' ? 'top: 0; left: 0' : 'bottom: 0; right: 0'};
-
-  ${StyledButton} {
-    margin: 5px;
-    ${({ $side }) => ($side === 'left' ? 'margin-right: 0' : 'margin-left: 0')};
-  }
-`
-
-const StyledGridInput = styled(LazyChangeInput)<{
-  $color: ColorInstance
-  $isHighlighted?: boolean
-}>`
-  width: 100%;
-  height: 100%;
-  outline: 1px solid rgba(0, 0, 0, 0.5);
-  border: none;
-  padding: 0;
-  background: ${({ $color, $isHighlighted }) =>
-    $isHighlighted
-      ? Color($color).lightness(90).hsl().string()
-      : Color($color).lightness(75).hsl().string()};
-  font-size: 20px;
-  text-align: center;
-
-  &:focus {
-    outline: 1px solid black;
-    box-shadow: 0 0 5px black inset;
-    z-index: 100;
-  }
-`
-
-const StyledGridControlsContainer = styled.div`
-  position: absolute;
-  user-select: none;
-  touch-action: none;
-
-  & > * {
-    z-index: 1001; // Above StyledGridInfo
-  }
 `
 
 const StyledGridContainer = styled.div<{
@@ -2908,99 +1435,6 @@ const StyledGridContainer = styled.div<{
   }
 `
 
-const StyledId = styled.div<{ $color: ColorInstance; $disabled?: boolean }>`
-  flex-shrink: 0;
-  font-family: var(--font-mono);
-  font-weight: 700;
-  font-size: 11px;
-  letter-spacing: 0.02em;
-  background: ${({ $color }) =>
-    Color($color).lightness(52).hsl().string() || '#333'};
-  color: #0a0d12;
-  padding: 4px 0;
-  border-radius: var(--r-sm);
-  width: 2.6em;
-  text-align: center;
-  cursor: ${({ $disabled }) => ($disabled ? 'normal' : 'grab')};
-`
-
-const StyledStreamLine = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 7px 8px;
-  border-radius: var(--r-sm);
-  border: 1px solid transparent;
-  font-size: 13px;
-  cursor: default;
-
-  &:hover {
-    background: var(--surface-2);
-    border-color: var(--border);
-  }
-
-  & > div {
-    min-width: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    color: var(--text);
-  }
-
-  svg {
-    height: 100%;
-  }
-`
-
-function CreateInviteInput({
-  onCreateInvite,
-}: {
-  onCreateInvite: (invite: { name: string; role: StreamwallRole }) => void
-}) {
-  const [inviteName, setInviteName] = useState('')
-  const [inviteRole, setInviteRole] = useState('operator')
-  const handleChangeName = useCallback<JSX.InputEventHandler<HTMLInputElement>>(
-    (ev) => {
-      setInviteName(ev.currentTarget.value)
-    },
-    [setInviteName],
-  )
-  const handleChangeRole = useCallback<
-    JSX.InputEventHandler<HTMLSelectElement>
-  >(
-    (ev) => {
-      setInviteRole(ev.currentTarget.value)
-    },
-    [setInviteRole],
-  )
-  const handleSubmit = useCallback<JSX.SubmitEventHandler<HTMLFormElement>>(
-    (ev) => {
-      ev.preventDefault()
-      setInviteName('')
-      setInviteRole('operator')
-      onCreateInvite({ name: inviteName, role: inviteRole as StreamwallRole }) // TODO: validate
-    },
-    [onCreateInvite, inviteName, inviteRole],
-  )
-  return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <input
-          onChange={handleChangeName}
-          placeholder="Name"
-          value={inviteName}
-        />
-        <select onChange={handleChangeRole} value={inviteRole}>
-          <option value="admin">admin</option>
-          <option value="operator">operator</option>
-          <option value="monitor">monitor</option>
-        </select>
-        <button type="submit">create invite</button>
-      </form>
-    </div>
-  )
-}
-
 const StyledNewInviteBox = styled.div`
   display: block;
   padding: 10px 12px;
@@ -3011,28 +1445,6 @@ const StyledNewInviteBox = styled.div`
   color: var(--text);
   font-size: 13px;
 `
-
-function AuthTokenLine({
-  id,
-  role,
-  name,
-  onDelete,
-}: {
-  id: string
-  role: StreamwallRole
-  name: string
-  onDelete: (id: string) => void
-}) {
-  const handleDeleteClick = useCallback(() => {
-    onDelete(id)
-  }, [id, onDelete])
-  return (
-    <div>
-      <strong>{name}</strong>: {role}{' '}
-      <button onClick={handleDeleteClick}>revoke</button>
-    </div>
-  )
-}
 
 const StyledStatusBar = styled.div`
   display: flex;
